Package library files on root
Not ideal but needed for the button namespace
