+++ .gitignore
@@ -20,3 +20,6 @@ dist
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+buttons.js
+react-pico-8.js
