+++ package.json
@@ -2,10 +2,11 @@
   "name": "react-pico-8",
   "version": "2.2.0",
   "description": "",
-  "main": "dist/react-pico-8.min.js",
+  "main": "react-pico-8.js",
   "src": "src/pico-8.js",
   "files": [
-    "dist",
+    "react-pico-8.js",
+    "buttons.js",
     "License.txt",
     "package.json",
     "README.md"
@@ -14,7 +15,7 @@
     "start": "yarn watch",
     "build": "NODE_ENV=production webpack",
     "watch": "NODE_ENV=development webpack --watch",
-    "clean": "rm -rf dist",
+    "clean": "rm react-pico-8.js buttons.js",
     "test": "echo \"No tests \" && exit 0"
   },
   "repository": {
